🛠️ Chore : try manual ssg conf
